+++ src/System/Common/Process/ProcessMemory.hpp
@@ -5,7 +5,7 @@ class CProcessMemory
 {
 public:
     CProcessMemory(void);
-    virtual void* Alloc(uint32 size);
+    virtual void* Alloc(size_t size);
     virtual void Free(void* mem);
     virtual ~CProcessMemory(void);
 };
